fix: Refactor footer
