+++ src/apps/front-office/design-system/layouts/Footer/FooterContactInfo/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FooterContactInfo";
