+++ src/apps/front-office/design-system/layouts/Footer/FooterLinks/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FooterLinks";
