+++ src/apps/front-office/design-system/layouts/Footer/FooterSocialMedia/FooterSocialMedia.tsx
@@ -1,0 +1,34 @@
+import {
+  facebook,
+  instagram,
+  pinterest,
+  twitter,
+  youtube,
+} from "assets/footerImages";
+
+const socialIcons = [
+  { src: facebook, alt: "facebook" },
+  { src: twitter, alt: "twitter" },
+  { src: instagram, alt: "instagram" },
+  { src: pinterest, alt: "pinterest" },
+  { src: youtube, alt: "youtube" },
+];
+
+export default function FooterSocialMedia() {
+  return (
+    <div className="text-center lg:text-left">
+      <div className="flex justify-center lg:justify-start gap-2 mb-2">
+        <h6 className="text-[#253D4E] font-bold leading-7">Follow Us</h6>
+        {socialIcons.map(({ src, alt }) => (
+          <img
+            key={alt}
+            src={src}
+            alt={alt}
+            className="w-[25px] bg-customGreen rounded-full cursor-pointer"
+          />
+        ))}
+      </div>
+      <p className="text-[14px]">Up to 15% discount on your first subscribe</p>
+    </div>
+  );
+}
