+++ src/apps/front-office/design-system/layouts/Footer/FooterSocialMedia/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FooterSocialMedia";
