+++ src/shared/assets/footerImages.ts
@@ -1,0 +1,29 @@
+import appleIcon from "./images/apple-icon-store.jpg";
+import googleIcon from "./images/google-icon-play.jpg";
+import iconClock from "./images/icon-clock.svg";
+import iconContact from "./images/icon-contact.svg";
+import iconEmail from "./images/icon-email.svg";
+import facebook from "./images/icon-facebook-white.svg";
+import instagram from "./images/icon-instagram-white.svg";
+import iconLocation from "./images/icon-location.svg";
+import pinterest from "./images/icon-pinterest-white.svg";
+import twitter from "./images/icon-twitter-white.svg";
+import youtube from "./images/icon-youtube-white.svg";
+import paymentMethod from "./images/payment-method.png";
+import phoneCall from "./images/phone-call.svg";
+
+export {
+  appleIcon,
+  facebook,
+  googleIcon,
+  iconClock,
+  iconContact,
+  iconEmail,
+  iconLocation,
+  instagram,
+  paymentMethod,
+  phoneCall,
+  pinterest,
+  twitter,
+  youtube,
+};
