+++ src/shared/localization/index.json
@@ -1,4 +1,8 @@
 {
+  "legalNotice": {
+    "en": " HTML E-commerce Template",
+    "ar": "قالب التجارة الإلكترونية HTML"
+  },
   "test": {
     "en": "Test",
     "ar": "التجربة"
